+++ src/django_upgrade/fixers/on_delete.py
@@ -88,7 +88,8 @@ def visit_Call(
         and not any(isinstance(arg, ast.Starred) for arg in node.args)
         and all(kw.arg != "on_delete" for kw in node.keywords)
     ):
-        should_update_import[state] = not models_imported
+        if not models_imported:
+            should_update_import[state] = True
         yield (
             ast_start_offset(node),
             partial(